fork update

+++ src/events/topics/one-word.js
@@ -1,12 +1,10 @@
 module.exports = {
-    event: "message"
+    event: "cleanMessage"
 }
 
 module.exports.call = async (bot, msg) => {
     if(msg.channel.type == 'dm') return;
     if(msg.channel.topicSetting('one-word')){
-        if(msg.content.match(/\s/g)){
-            msg.delete().catch(()=>{})
-        }
+        if(msg.content.match(/\s/g)) msg.delete().catch(()=>{});
     }
 }
